feat(82): final - with description
Dmn engine now has persistence for DecisionDefinitions and has few queries to get them from db. ( tests included )
Sqlc overrides were removed for clarity. And for now types are cast from int64.
Fixed FindProcessDefinitionsById query since there were issues found with it during development.

+++ internal/cluster/node.go
@@ -386,7 +386,7 @@ func (node *ZenNode) GetProcessDefinitions(ctx context.Context) ([]proto.Process
 	for _, def := range definitions {
 		resp = append(resp, proto.ProcessDefinition{
 			Key:        def.Key,
-			Version:    def.Version,
+			Version:    int32(def.Version),
 			ProcessId:  def.BpmnProcessID,
 			Definition: []byte(def.BpmnData),
 		})
@@ -406,7 +406,7 @@ func (node *ZenNode) GetLatestProcessDefinition(ctx context.Context, processId s
 	}
 	return proto.ProcessDefinition{
 		Key:        def.Key,
-		Version:    def.Version,
+		Version:    int32(def.Version),
 		ProcessId:  def.BpmnProcessID,
 		Definition: []byte(def.BpmnData),
 	}, nil
@@ -424,7 +424,7 @@ func (node *ZenNode) GetProcessDefinition(ctx context.Context, key int64) (proto
 	}
 	return proto.ProcessDefinition{
 		Key:        def.Key,
-		Version:    def.Version,
+		Version:    int32(def.Version),
 		ProcessId:  def.BpmnProcessID,
 		Definition: []byte(def.BpmnData),
 	}, nil
